requirements.txt
paho-mqtt==1.3.1
pytest==3.6.3
sdnotify==0.3.2
